\name{auckpop2006}
\docType{data}
\alias{auckpop2006}
\title{Ethnic groups in the Auckland Metropolitan Areas}
\description{
This data set contains the digital boundaries of 297 census tracts in the Auckland metropolitan areas.
}
\usage{
data(auckpop2006)
}
\format{
The data set \code{auckpop2006} is a class of \code{SpatialPolygonsDataFrame}, and the data slot contains population counts by census tracts (rows) and ethnic groups (columns). Each column indicates the following information:

\sQuote{EUROPEAN} The number of Europeans who identified with one or more European ethnicities in the 2006 census. The largest ethnicity among European people is New Zealand European, followed by English, Dutch, Australian, South African, Scottish and Irish.

\sQuote{MAORI} The number of Maori people who identified with the Maori ethnic group in the 2006 census. 

\sQuote{PACIFIC} The number of Pacific peoples who identified with one or more Pacific ethnicities in the 2006 census. The seven largest ethnicities among Pacific peoples are Cook Island Maori, Fijian, Niuean, Samoan, Tokelauan, Tongan and Tuvaluan peoples.

\sQuote{ASIAN} The number of Asian people who identified with one or more Asian ethnicities in the 2006 census. The largest ethnicity among Asian people is Chinese, followed by Indian, Korean, Filipino, Japanese, Sri Lankan, Cambodian and Thai. 

\sQuote{MELAA} The number of people who were represented in MELAA in the 2006 census. The MELAA ethnic group includes people of Middle Eastern, African and Latin American ethnicities.

\sQuote{OTHER} The number of people who belonged to other groups that do not fit into the five other groupings.
}
\source{
\href{http://www.stats.govt.nz/Census/about-2006-census.aspx}{Statistics New Zealand} and licensed by Statistics New Zealand for re-use under the \href{http://creativecommons.org/licenses/by/3.0/nz/deed.en}{Creative Commons Attribution 3.0 New Zealand licence}.
}
\keyword{datasets}
