% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.R
\name{dissim}
\alias{dissim}
\title{Index of Dissimilarity}
\usage{
dissim(x, data, adjust = FALSE, use_queen = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{An optional [sf::sf] object with polygon geometry.}

\item{data}{A numeric matrix or data frame with two columns that represent
mutually exclusive population groups (e.g., Asians and non-Asians). If
more than two columns are given, only the first two will be used for
computing the index.}

\item{adjust}{Logical. If TRUE, and if `x` is given, the index of
dissimilarity is adjusted according to the suggestions of Morrill (1991)
and Wong (1993), depending on packages installed on the system. See
Details for more information. Ignored if `x` is not given.}

\item{use_queen}{Logical. If TRUE, the queen contiguity will be used.}

\item{verbose}{Logical. If TRUE, print the current stage of the computation
and time spent on each job to the screen.}
}
\value{
A list containing:
  \item{d}{Index of dissimilarity.}
  \item{dm}{Index of dissimilarity adjusted according to Morrill (1991).
    NA if not calculated.}
  \item{dw}{Index of dissimilarity adjusted according to Wong (1993).
    NA if not calculated.}
  \item{ds}{Index of dissimilarity adjusted according to Wong (1993).
    NA if not calculated.}
}
\description{
Calculates the index of dissimilarity proposed by Duncan and Duncan (1955).
If `x` is given and `adjust = TRUE`, the index is adjusted to reflect the
spatial distribution of population.
}
\details{
`dissim` calculates the index of dissimilarity for `data`. If `data` is
missing, it attempts to extract the data from `x`. If `x` is not given, or
if it is not a valid [sf::sf] object, the function stops with an error.

When `x` is given and `adjust` is set to TRUE, the index is adjusted
according to the suggestions of Morrill (1991) and Wong (1993). This
automatic adjustment uses [spdep] for Morrill's D(adj) and [terra] for
Wong's D(w) and D(s), when those packages are available.
}
\examples{
if (requireNamespace("spdep", quietly = TRUE)) {
  # uses the idealised landscapes in 'segdata'
  data(segdata)
  geom <- sf::st_polygon(list(rbind(c(0,0), c(10,0), c(10,10), c(0,0)))) |>
    sf::st_make_grid(cellsize = 1)
  grd_sf <- sf::st_sf(geom)
  grd.nb <- spdep::nb2mat(spdep::poly2nb(grd_sf, queen = FALSE), style = "B")
  grd.nb <- grd.nb / sum(grd.nb)

  d <- list(); m <- list()
  for (i in 1:8) {
    idx <- 2 * i
    d[[i]] <- dissim(data = segdata[,(idx-1):idx])
    m[[i]] <- dissim(grd_sf, segdata[,(idx-1):idx], adjust = TRUE)
  }
}

}
\references{
Duncan, O. D., & Duncan, B. (1955). A methodological analysis of segregation
  indexes. *American Sociological Review*, **20**, 210-217.

Morrill, R. L. (1991). On the measure of geographic segregation.
  *Geography Research Forum*, **11**, 25-36.

Wong, D. W. S. (1993). Spatial indices of segregation. *Urban Studies*,
  **30**, 559-572.
}
\seealso{
[spseg()]
}
