% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isp.R
\name{isp}
\alias{isp}
\title{Index of Spatial Proximity}
\usage{
isp(x, data, nb, fun, verbose = FALSE, ...)
}
\arguments{
\item{x}{A numeric matrix or data frame with coordinates (each row is a
point), or an object of class [sf::sf].}

\item{data}{An object of class `matrix`, or one that can be coerced to
that class. The number of rows in `data` should equal the number of
geographic units in `x`, and the number of columns should be greater
than one (i.e., at least two population groups are required). This can
be missing if `x` has a data frame attached to it.}

\item{nb}{An optional `matrix` object indicating the distances between
the geographic units.}

\item{fun}{A function for the calculation of proximity. The function
should take a numeric vector as an argument (distance) and return a
vector of the same length (proximity). If this is not specified, a
negative exponential function is used by default.}

\item{verbose}{Logical. If TRUE, print the current stage of the
computation and time spent on each job to the screen.}

\item{...}{Optional arguments to be passed to [stats::dist()] when
calculating the distances between the geographic units in `x`. Ignored
if `nb` is given. See [stats::dist()] for available options.}
}
\value{
A single numeric value indicating the degree of segregation; a
  value of 1 indicates absence of segregation, and values greater than 1.0
  indicate clustering. If the index value is less than one, it indicates
  an unusual form of segregation (i.e., people live closer to other
  population groups).
}
\description{
Computes the index of spatial proximity developed by White (1983). This
measure estimates the level of clustering by comparing the average distance
between members of the same group with that between all individuals
(regardless of the groups to which they belong). The results may change
drastically depending on the definition of distance.
}
\details{
`nb` must be a square matrix (i.e., one that has the same number of rows
and columns) but does not have to be symmetric. If `nb` is not given,
`isp` attempts to create a distance matrix of `x` using the function
[stats::dist()] and use it as `nb`. The optional arguments in `...` will
be passed to [stats::dist()].
}
\examples{
# uses the idealised landscapes in 'segdata'
data(segdata)
geom <- sf::st_polygon(list(rbind(c(0,0), c(10,0), c(10,10), c(0,0)))) |>
  sf::st_make_grid(cellsize = 1)
grd_sf <- sf::st_sf(geom)

d <- rep(NA, 8) # index of dissimilarity
p <- rep(NA, 8) # index of spatial proximity
for (i in 1:8) {
  idx <- 2 * i
  d[i] <- dissim(data = segdata[,(idx-1):idx])$d
  p[i] <- isp(grd_sf, data = segdata[,(idx-1):idx])
}

if (requireNamespace("ggplot2", quietly = TRUE)) {
  plot_data <- do.call(
    rbind,
    lapply(1:8, function(i) {
      idx <- 2 * i - 1
      data.frame(
        pattern = paste0(
          LETTERS[i], ": D = ", round(d[i], 2), ", P = ", round(p[i], 2)
        ),
        group_share = segdata[, idx],
        geometry = sf::st_geometry(grd_sf)
      )
    })
  )
  plot_data <- sf::st_sf(plot_data)

  ggplot2::ggplot(plot_data) +
    ggplot2::geom_sf(ggplot2::aes(fill = group_share), color = "white") +
    ggplot2::facet_wrap(~ pattern) +
    ggplot2::scale_fill_viridis_c() +
    ggplot2::theme_void()
}

}
\references{
White, M. J. (1983). The measurement of spatial segregation.
  *The American Journal of Sociology*, **88**, 1008-1018.
}
\seealso{
[dissim()], [stats::dist()]
}
